Enhance debt and expense management: update data models to use Int for amounts, implement SwipeToDeleteContainer for item deletion, and improve UI components with error handling and keyboard options.

+++ app/src/main/java/com/example/subsense/core/model/Expense.kt
@@ -8,7 +8,7 @@ import com.example.subsense.setting.data.model.RecurringPattern
 data class Expense(
     @PrimaryKey(autoGenerate = true)
     val id: Long,
-    val amount: Double?,
+    val amount: Int?,
     val category: ExpenseCategory,
     val date: Long = System.currentTimeMillis(), // timestamp in milliseconds
     val note: String?,
