+++ app/src/main/java/com/example/subsense/expense/presentation/manager/state/ExpenseState.kt
@@ -4,7 +4,7 @@ import com.example.subsense.core.model.Budget
 import com.example.subsense.core.model.Expense
 
 data class ExpenseState(
-    val totalSpend: Long = 0,
+    val totalSpend: Int = 0,
     val expenses: List<Expense> = emptyList(),
     val budgetLimit: List<Budget> = emptyList(),
     val isLoading: Boolean = false
